% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_Laplace}
\alias{get_Laplace}
\title{Get the Laplace density}
\usage{
get_Laplace(x, lambda, mu = 0)
}
\arguments{
\item{x}{A number.}

\item{lambda}{A number for rate parameter (1/scale).}

\item{mu}{A number for location parameter (default = 0).}
}
\value{
A number.
}
\description{
Get the Laplace density
}
\examples{
get_Laplace(0, 1)
get_Laplace(0, 1, 2)
}
