% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_logit}
\alias{get_logit}
\title{Get the logit values.}
\usage{
get_logit(x)
}
\arguments{
\item{x}{A number or a list.}
}
\value{
A number or a list.
}
\description{
Get the logit values.
}
\examples{
get_logit(0)
}
